More changes to terrible code.
* Use size_t instead of off_t for Object::MMap.
* Change Eport to descend from Object.
* Add Truncate to Object.
* Stub clock_gettime syscall.
* Add evf_open, ftruncate, shm_open syscalls.
* Implement evf_delete.
* Add EventFlag, SharedMemory objects.

+++ uplift/src/devices/console_device.cpp
@@ -45,7 +45,7 @@ SCERR ConsoleDevice::IOControl(uint32_t request, void* argp)
   return SCERR::eNOSYS;
 }
 
-SCERR ConsoleDevice::MMap(void* addr, size_t len, int prot, int flags, off_t offset, void*& allocation)
+SCERR ConsoleDevice::MMap(void* addr, size_t len, int prot, int flags, size_t offset, void*& allocation)
 {
   assert_always();
   return SCERR::eNOSYS;
