+++ uplift/src/devices/deci_tty_device.cpp
@@ -45,7 +45,7 @@ SCERR DeciTTYDevice::IOControl(uint32_t request, void* argp)
   return SCERR::eNOSYS;
 }
 
-SCERR DeciTTYDevice::MMap(void* addr, size_t len, int prot, int flags, off_t offset, void*& allocation)
+SCERR DeciTTYDevice::MMap(void* addr, size_t len, int prot, int flags, size_t offset, void*& allocation)
 {
   assert_always();
   return SCERR::eNOSYS;
