+++ uplift/src/devices/dipsw_device.cpp
@@ -69,7 +69,7 @@ SCERR DipswDevice::IOControl(uint32_t request, void* argp)
   return SCERR::eINVAL;
 }
 
-SCERR DipswDevice::MMap(void* addr, size_t len, int prot, int flags, off_t offset, void*& allocation)
+SCERR DipswDevice::MMap(void* addr, size_t len, int prot, int flags, size_t offset, void*& allocation)
 {
   assert_always();
   return SCERR::eNOSYS;
