+++ uplift/src/devices/direct_memory_device.cpp
@@ -79,7 +79,7 @@ SCERR DirectMemoryDevice::IOControl(uint32_t request, void* argp)
   return SCERR::eNOSYS;
 }
 
-SCERR DirectMemoryDevice::MMap(void* addr, size_t len, int prot, int flags, off_t offset, void*& allocation)
+SCERR DirectMemoryDevice::MMap(void* addr, size_t len, int prot, int flags, size_t offset, void*& allocation)
 {
   assert_always();
   return SCERR::eNOSYS;
