+++ uplift/src/sockets/internet_socket.cpp
@@ -145,7 +145,7 @@ SCERR InternetSocket::IOControl(uint32_t request, void* argp)
   return SCERR::eNODEV;
 }
 
-SCERR InternetSocket::MMap(void* addr, size_t len, int prot, int flags, off_t offset, void*& allocation)
+SCERR InternetSocket::MMap(void* addr, size_t len, int prot, int flags, size_t offset, void*& allocation)
 {
   assert_always();
   return SCERR::eNODEV;
